ktap: add compiler(tools/ktap/kp_[lex|parse].[c|h])
ktap compiler is based on luajit, origin forked from lua.

The compiler use one pass compilation, it is very fast, and
the generated binary is very small.

[root@localhost ktap]# ll -h ktap
-rwxr-xr-x. 1 root root 83K Mar 27 06:09 ktap

The compiler is easy to hack, the code is very easy to understand.

Note that some bytecode in luajit is not used by ktap, it could
be remove in future.

More compiler hack soon:
1). aggregation
	@name[keys] = aggfunction(args)

2). multi-key table
	var s = {}
	s[key1, key2, key3] = value

3). C structure access

Signed-off-by: Jovi Zhangwei <jovi.zhangwei@gmail.com>